Added many features.
Mana, Health, UI for both, new cards and attacks. Reworked Tags because it was stupid to make them resources. They are now an enum in the card class. I refactored the cards to make them less restricted in implementation of the abstract card class. I figured out tilemaps and used it for a simple field. Working on the UI for the deck. Functional for now, but not visible.

I'm building this with controller in mind, but making it work with k/m rn.